refactor: move functions for gamer answers to own module manage_answer.js
